Bind to uniffi using proc macros
Proc macros let rust-analyzer check that types are compatible
at rust compile time, and thus let downstream projects more easily
depend on bitcoin-ffi.

The `Network` enum is defined in the UDL file, but we need to define it
in rust with `uniffi::Enum` in order to check UniFFI bindings at rust
compile time.

"It is permitted to use this [`uniffi::Enum`] macro on a type that is
also defined in the UDL file as long as the two definitions are equal in
the names and ordering of variants and variant fields"

see: https://mozilla.github.io/uniffi-rs/latest/proc_macro/index.html#the-uniffienum-derive